Deduplicate most-read results
Moves hydrated title deduplication from onthisday to a util function where
it can be reused for most-read.

Alternate solution to https://gerrit.wikimedia.org/r/#/c/mediawiki/services/mobileapps/+/434806/.

Bug: https://phabricator.wikimedia.org/T195390
