+++ lib/mwUtil.js
@@ -563,4 +563,20 @@ mwUtil.constructRegex = (list) => {
     return regex;
 };
 
+/**
+ * Remove entries with duplicate 'title' values from an array of objects
+ * @param {?Array} arr Array of page/article objects
+ * @return {?Array} arr deduplicated by 'title'
+ */
+mwUtil.removeDuplicateTitles = (arr) => {
+    const titles = {};
+    return arr.filter((item) => {
+        if (titles[item.title]) {
+            return false;
+        }
+        titles[item.title] = true;
+        return true;
+    });
+};
+
 module.exports = mwUtil;
